Updated CI specs to RubySpec ee468c6b.

+++ spec/ruby/library/bigdecimal/shared/modulo.rb
@@ -67,7 +67,7 @@ describe :bigdecimal_modulo, :shared => true do
     @two.send(@method, 2.0).should == 0.0
     @one.send(@method, 2.0).should == 1.0
     res = @two.send(@method, 5.0)
-    res.kind_of?(Float).should == true
+    res.kind_of?(BigDecimal).should == true
   end
 
   it "returns NaN if NaN is involved" do
